requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
CacheControl==0.14.4
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
cryptography==46.0.3
distro==1.9.0
fastapi==0.128.0
firebase_admin==7.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
msgpack==1.1.2
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.40.0
websockets==15.0.1
